fix: Rename advisor.rs to core.rs to avoid module_inception clippy warning
